Replace answer button radio inputs with buttons for better keyboard support
One of the quiz options is showing the correct answer when an answer is selected. This causes issues with radio inputs because navigating to a radio input with the keyboard will automatically select it, at which point the correct answer will be shown and the rest of the radio inputs will be disabled. This means it's not possible to access the answers using the keyboard before selecting one.